+++ templates/admin-panel/template/client/__tests__/unit/components/routes/server-info/assets-panel.js
@@ -1,0 +1,7 @@
+import AssetsPanel  from '../../../../../src/components/routes/server-info/assets-panel.sfc';
+import renderer from 'sham-ui-test-helpers';
+
+it( 'renders correctly', () => {
+    const meta = renderer( AssetsPanel, {} );
+    expect( meta.toJSON() ).toMatchSnapshot();
+} );