+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Fix a tests
 * Make `settings` page in admin panel lazy
 * Refactoring session service in admin-panel
+* Refactoring `server-info` page component in admin-panel
 
 ## 4.3.0
 * Add template for `styled-component` [#3](https://github.com/sham-ui/sham-ui-cli/issues/3)